fix: move doc mapping tests to doc mapping module
